Pre-alpha SeaBattle Bot
IS NOT READY TO RELEASE!!
About 50% of the final program is completed.

Currently implemented:
* Ability to start chatting with a bot.
* Authorization of users by ID (not saved when the bot is disabled).
* Greeting (without commands "help" and "rules")
* Automatic placement of player and AI ships.
* Ability to change the display of player and AI fields.
* Ability to enter the coordinates of the shot by the player (with checking the validity of the shot)
* Ability to reselect the coordinates of the shot with a failed check.
In the process of implementing the functionality of a full-fledged battle, at the moment the mechanics of a shot and a hit are implemented.

Attention!! the project is poorly structured, its template is expected to undergo significant changes later!