Configuration change!
All handlers that work with User class data have been rewritten using the new decorator.

Minor changes have been made to the code to improve readability.

P.S.
I noticed that the process of randomly choosing coordinates for a shot is not optimized. When there are practically no free cells on the field, we must get the coordinates of the remaining cells randomly, which can take a lot of time. TO THE FIX!!!